compiling with fatfs r0.10c

+++ kitsune/audiohelper.c
@@ -140,9 +140,9 @@ uint8_t InitFile(Filedata_t * pfiledata) {
 
 }
 
-uint8_t WriteToFile(Filedata_t * pfiledata,const WORD bytes_to_write,const uint8_t * const_ptr_samples_bytes) {
-	WORD bytes = 0;
-	WORD bytes_written = 0;
+uint8_t WriteToFile(Filedata_t * pfiledata,const UINT bytes_to_write,const uint8_t * const_ptr_samples_bytes) {
+	UINT bytes = 0;
+	UINT bytes_written = 0;
 	FRESULT res;
 	uint8_t ret = 1;
