+++ kitsune/fileuploadertask.c
@@ -98,8 +98,8 @@ bool encode_file (pb_ostream_t * stream, const pb_field_t * field,void * const *
 	uint8_t readbuf[256];
 
 	DWORD bytes_to_be_read;
-	WORD read_size;
-	WORD bytes_read;
+	UINT read_size;
+	UINT bytes_read;
 
 	bool success = false;
